ppc: Apple Partition Map write support + Tabby disk installer
Everything needed to partition, format, and install Tabby onto a real
disk from the GUI Installer. Proven end-to-end: the first fresh install
of Tabby booting from an iBook G3's internal IDE disk.

- apple partitioning module: write support (initialize / create_child /
  delete_child, supported-operations flags, get_partitionable_spaces,
  validate_*); expose child type/name; skip the self map entry from the
  framework child tree (kept DriveSetup/Installer CommitModifications from
  failing). Fix a 32-bit multiply overflow in the map size helpers
  (blockSize*blockCount was computed in int, wrong for disks/partitions
  > 4 GB - the Installer saw ~320 KB free on an 18 GB disk); compute in
  64-bit (apple.h Start()/Size(), apple.cpp content_size). Adds
  PartitionLocker + apple_write_support.cpp.
- disk_systems/apple: new userland AppleDiskAddOn so DriveSetup/Installer
  can create an Apple map and child partitions.
- bin/tabby_install: CLI that partitions (APM + Apple_HFS loader +
  Haiku_BFS) and formats the BFS; erases existing partitions first so a
  re-run on an already-partitioned or corrupt disk starts from a clean
  slate (Initialize alone did not remove pre-existing children).
- makebootable (openfirmware): write the loader onto the target's
  Apple_HFS partition (accepts a raw-disk path and a shipped loader blob),
  idempotent.
- Installer: "Set up disk for Tabby" tool + boot-sector write wiring.

+++ src/add-ons/kernel/partitioning_systems/apple/apple.h
@@ -57,8 +57,8 @@ struct apple_partition_map {
 
 	int32 StartBlock() { return B_BENDIAN_TO_HOST_INT32(start); }
 	int32 BlockCount() { return B_BENDIAN_TO_HOST_INT32(size); }
-	uint64 Start(apple_driver_descriptor &descriptor) { return StartBlock() * descriptor.BlockSize(); }
-	uint64 Size(apple_driver_descriptor &descriptor) { return BlockCount() * descriptor.BlockSize(); }
+	uint64 Start(apple_driver_descriptor &descriptor) { return (uint64)StartBlock() * descriptor.BlockSize(); }
+	uint64 Size(apple_driver_descriptor &descriptor) { return (uint64)BlockCount() * descriptor.BlockSize(); }
 
 	bool HasValidSignature();
 };
